connectionHealthCheck: rename to onQueryError and improve
This changes the logic to allow I/O action, and return a more obvious
enum for whether to keep the connection or drop it.

It then changes the default setting to drop the connection on *any*
ClientError, not match on the libpq message. If there is a
`ClientError` it means that the connection is bad with a high
probability.

libpq also has a for a `reset :: Connection -> IO ()` function, but it
doesn’t give any promises on whether that works or not, except for
“this might repair the connection”, and our pool will create another
connection anyway (which might be a bit more work, so we leave it up
to clients whether they want to use this and return `KeepConnection`
if successful …)

Since `onQueryError` is now in IO, we also catch any errors there and
drop the connection if something happens.