Reorder & improve documentation.
The exports of the module are sorted in a way that they make more
sense to a user of the module.

A few docstrings are added, others clarified.

Especially what errors mean to the retaining of connections in the
pool is clarified.

It is painfully obvious to me now that `Hasql.ClientError` is the same
kind of error as `ConnectionError`, so a future version of Hasql.Pool
might opt to unify these two. Of course that would mean re-shuffling
the error returned by hasql, e.g. by putting them into one enum
instead of two nested ones.